Fix TokenExplorerModal currency formatting issues
